Stroop test vol 2
I try to create here a sort of reaction game, inspired by Daniel Kahneman in his book "Thinking fast and slow". 
My goal is to create two kinds of phases in this game. First, the participants see a task "Pick object that is on the left/right side of the page. Then they see two objects, one on the right side and the other on the left side. Both of those objects have a word (left/right) written on in the front. Participants have to click the right one as quickly as possible. Perhaps, this could be repeated like 10 times. At the same time, there is a clock that measures how long does this take. If there are errors, then the test is failed. 
The second phase is similar one but now texts (left/right) are randomly picked.